Renamed classes and updates.
