+++ src/main/scala-2.11/com/bwsw/tstreams/coordination/pubsub/ProducerToSubscriberNotifier.scala
@@ -1,8 +1,7 @@
 package com.bwsw.tstreams.coordination.pubsub
 
 import java.net.InetSocketAddress
-
-import com.bwsw.tstreams.common.zkservice.ZkService
+import com.bwsw.tstreams.common.ZookeeperDLMService
 import com.bwsw.tstreams.coordination.pubsub.messages.ProducerTopicMessage
 import com.bwsw.tstreams.coordination.pubsub.publisher.Broadcaster
 import org.apache.zookeeper.{WatchedEvent, Watcher}
@@ -24,7 +23,7 @@ class ProducerToSubscriberNotifier(prefix: String,
                                    zkSessionTimeout: Int,
                                    zkConnectionTimeout: Int) {
 
-  private val zkService = new ZkService(prefix, zkHosts, zkSessionTimeout, zkConnectionTimeout)
+  private val zkService = new ZookeeperDLMService(prefix, zkHosts, zkSessionTimeout, zkConnectionTimeout)
   private val broadcaster = new Broadcaster
 
   /**
