+++ src/main/scala-2.11/com/bwsw/tstreams/coordination/pubsub/SubscriberCoordinator.scala
@@ -1,8 +1,7 @@
 package com.bwsw.tstreams.coordination.pubsub
 
 import java.net.InetSocketAddress
-
-import com.bwsw.tstreams.common.zkservice.ZkService
+import com.bwsw.tstreams.common.ZookeeperDLMService
 import com.bwsw.tstreams.coordination.pubsub.listener.SubscriberListener
 import com.bwsw.tstreams.coordination.pubsub.messages.ProducerTopicMessage
 import org.apache.zookeeper.{CreateMode, KeeperException}
@@ -23,7 +22,7 @@ class SubscriberCoordinator(agentAddress: String,
 
   private val logger = LoggerFactory.getLogger(this.getClass)
   private val SYNCHRONIZE_LIMIT = 60
-  private val zkService = new ZkService(zkRootPrefix, zkHosts, zkSessionTimeout, zkConnectionTimeout)
+  private val zkService = new ZookeeperDLMService(zkRootPrefix, zkHosts, zkSessionTimeout, zkConnectionTimeout)
   private val (_, port) = getHostPort(agentAddress)
   private val listener: SubscriberListener = new SubscriberListener(port)
   private var isFinished = false
