+++ src/main/scala-2.11/com/bwsw/tstreams/coordination/transactions/transport/impl/client/TransactionStateUpdateClient.scala
@@ -2,9 +2,8 @@ package com.bwsw.tstreams.coordination.transactions.transport.impl.client
 
 import java.io.{BufferedReader, IOException, InputStreamReader}
 import java.net.{Socket, SocketTimeoutException}
-
-import com.bwsw.tstreams.common.serializer.TStreamsSerializer
-import com.bwsw.tstreams.common.serializer.TStreamsSerializer.TStreamsSerializerException
+import com.bwsw.tstreams.common.TStreamsSerializer
+import TStreamsSerializer.TStreamsSerializerException
 import com.bwsw.tstreams.coordination.transactions.messages.IMessage
 import org.slf4j.LoggerFactory
