+++ src/main/scala-2.11/com/bwsw/tstreams/coordination/transactions/transport/impl/server/TransactionStateMessageChannelHandler.scala
@@ -1,9 +1,8 @@
 package com.bwsw.tstreams.coordination.transactions.transport.impl.server
 
 import java.util
-
-import com.bwsw.tstreams.common.serializer.TStreamsSerializer
-import com.bwsw.tstreams.common.serializer.TStreamsSerializer.TStreamsSerializerException
+import com.bwsw.tstreams.common.TStreamsSerializer
+import TStreamsSerializer.TStreamsSerializerException
 import com.bwsw.tstreams.coordination.transactions.messages.IMessage
 import io.netty.channel._
 import io.netty.handler.codec.{MessageToMessageDecoder, MessageToMessageEncoder}
